requirements.txt
SQLAlchemy
pandas
psycopg2-binary
polygon-api-client
flask